Fixign race condition in temporary file naming
Ok, this is super gross:

File names were being created using a request counter, that request
counter was set when a new request was encountered. The function that
created the file names (referencing that request counter) was being
called in the context of callbacks (various) so.... we effectively had a
race condition whereby creation of file names used for different
requests could actually result in the same filename... So in the case of
many concurrent requests forkulator just flat could return the wrong
info... whoops.